+++ derive_builder/examples/readme_example.rs
@@ -1,3 +1,5 @@
+#![cfg_attr(feature = "try_from", feature(try_from))]
+
 // NOTE: generate fully expanded version with `cargo expand`.
 //
 //       cargo expand --example readme_example