+++ derive_builder/tests/build_fn.rs
@@ -1,3 +1,4 @@
+#![cfg_attr(feature = "try_from", feature(try_from))]
 #[macro_use]
 extern crate derive_builder;
 
@@ -38,7 +39,7 @@ impl LoremBuilder {
 impl IpsumBuilder {
     /// This should be fine, because we renamed the generated build_fn.
     #[allow(dead_code)]
-    fn build(&self) -> Result<Self, String> {
+    fn build(&self) -> Result<Ipsum, String> {
         unimplemented!()
     }
 }