+++ derive_builder/tests/run-pass/attributes.rs
@@ -1,3 +1,6 @@
+#![allow(unused_features)]
+#![feature(try_from)]
+
 #[macro_use]
 extern crate derive_builder;