+++ derive_builder/tests/run-pass/custom_types.rs
@@ -1,4 +1,5 @@
-#![allow(dead_code)]
+#![allow(unused_features, dead_code)]
+#![feature(try_from)]
 
 #[macro_use]
 extern crate derive_builder;